+++ packages/dom-snapshot/test/util/hang.js
@@ -1,0 +1,6 @@
+module.exports = (req, _res, next) => {
+  console.log('xxx', req.url, req.method);
+  if (req.url !== '/hanging.css') {
+    next();
+  }
+};